rados export: better name mangling rules, fix test
Introduce a versioning scheme for name mangling, so that we can change
it in the future if we want to.

For names that require no mangling, don't do anything (don't add any
hash at the end, etc.) Make the hash 16 bytes long rather than 8 bytes.

Fix the unit test to take into account the new rules.

Signed-off-by: Colin McCabe <colin.mccabe@dreamhost.com>