mds: use XSYN state for rdlocks during EXCL
Move to XSYN state if we get an rdlock attempt from EXCL.  This means that
when there is an EXCL client doing buffered writes, we move to XSYN, which
stops the writer but does not require that data be flushed to disk.  This
vastly improves performance in cases where you have (perhaps many)
writer(s) in a directory and another client does stat or ls -al.

Signed-off-by: Sage Weil <sage@newdream.net>