ReplicatedPG,PG: update snap_collections on replica
Previously, snap_collections did not get updated on the replica.  As a
result, snap collections would not necessarily get trimmed when the
replica recieved and updated purged_snaps via a pg info from the
primary.  Now, the log entries in sub_op_modified are scanned to check
for any new snap collections.  The first and last snaps on a clone entry
are possibly new snap collections.

Signed-off-by: Samuel Just <samuel.just@dreamhost.com>