mds: make issue_caps from file_update_finish smarter
We do one funky thing in file_update_finish that only issues caps on a
single cap when max_size changes.  This is more commonly we see.  However,
if a lock changes state and we need to issue on the whole inode (for all
clients), avoid doing the cap-specific issue by checking the issue set.

Signed-off-by: Sage Weil <sage@newdream.net>