mds: fix transitions away from XSYN state
If we are in XSYN, we need to transition first to EXCL before moving on.
Simply going to the other unstable state isn't enough, because if gather
is zero at the end of the method we end up jumping directly to the wrong
state without passing EXCL.   Instead, call file_excl(), and continue only
if we got there immediately.

Signed-off-by: Sage Weil <sage@newdream.net>