mds: when prodding head inode on snapped behalf, avoid XSYN
If we try to rdlock a snapped inode in snap->sync state, we need to prod
the head inode into SYNC to get the data flushed.  Make sure we avoid
XSYN by passing as_anon=true to rdlock_start(head).

Also fix the waiting code.  If it succeeds, continue around the loop.  If
not, the rdlock_start will add the waiter for us--don't do it again!

Fixes workunit snaps/snaptest-snap-rm-cmp.sh.

Signed-off-by: Sage Weil <sage@newdream.net>