AnchorServer: overwrite old Anchor backpointers when proper.
Sometimes when we do an AnchorTable update, it's because the inode
in question got moved. However, if the inode had multiple references,
its Anchor wasn't removed by decrementing its count, and so the add
function simply noted that it already had the Anchor and returned.

This obviously wasn't the proper behavior in cases where the inode
was getting moved -- we want to update its back pointer! So do so.

Signed-off-by: Greg Farnum <gregory.farnum@dreamhost.com>