client: always use get_snap_context() accessor
There were a few places where we were using the member directly and not the
accessor, which rebuilds the internal value when needed.  This caused
inconsistent behavior based on whether debugging was enabled or not, since
we used the accessor to print the regenerated value.

Passes all snaps workunits with debugging off.

Signed-off-by: Sage Weil <sage@newdream.net>