C_GatherBuilder: add C_GatherBuilder::activate()
Add an activate() function that must be called before we call the
onfinish callback. This is especially important in multi-threaded
contexts, since otherwise if completions come in in the wrong order, we
may delete the C_Gather object right before calling new_sub on it!

Also delete rm_subs because it is redundant with sub_finish.

Finally, num_subs_created, num_subs_remaining are now methods on
C_GatherBuilder rather than C_Gather.

Signed-off-by: Colin McCabe <colin.mccabe@dreamhost.com>