mds: fix snaprealm split for dir inode
The snaprealm root directory inode belongs to the snaprealm, at least
currently.  This make split_at() consistent with the non-directory case
at the top of the method, and prevents a crash later down the line when we
try to tear down the parent snaprealm and we aren't part of it.

Fixes: #1238
Signed-off-by: Sage Weil <sage.weil@dreamhost.com>