client: only send one flushsnap once per mds session
This mirrors a kclient change a while back (e835124).

We only want to send one flushsnap cap message per MDS session:
 - it's a waste to send multiples
 - the mds will only reply to the first one

If the mds restarts we need to resend.

This fixes a hang where we send multiples, the first (and only) reply is
ignored (due to tid mismatch), and we are left with dangling references to
the inode and hang on umount.  (Reliably reproduced by running the full
snaps/ workunit directory.)

Fixes: #1239
Signed-off-by: Sage Weil <sage.weil@dreamhost.com>