context: implement complete()
finish() requires the caller to delete.  complete() does that for you by
calling finish() and then doing delete this.  Unless you overload it and
do something else.  This will allow us to make Contexts are are reusable,
for example, by overloading complete() instead of finish() and managing
the lifecycle in some other way.

Signed-off-by: Sage Weil <sage.weil@dreamhost.com>