mds: add mix->lock(2) state
There is a problem with the wrlocks and cross-mds renames:

 - master (dest auth, srci auth, srcdir replica) takes wrlock on srcdiri
 - something triggers a srcdiri lock, putting inest/ifile lock in mix->lock
   state
 - slave (srcdir auth) sends LOCKACK
 - master sends prepare_rename
 - slave (srcdir auth) does rename prepare, which modifies srcdir

Even though the master holds a wrlock on the srcdiri, the gather starts
immediately and the slave sends the LOCKACK before the master's wrlock is
released.

To fix this, we add a new mix->lock(2) state, and we do not start the
mix->lock gather from replicas until the local gather completes, _after_
the auth's wrlock is released.  This makes the master's wrlock sufficient
to ensure the prepare_rename on the slave is save.

This also works when the slave is the srci auth, since the gather won't
complete until the master releases its wrlock.  BUT, it does NOT work if a
third MDS is the srcdiri auth, since it can still gather from the slave
prior to the master releasing its wrlock.

Signed-off-by: Sage Weil <sage.weil@dreamhost.com>