mds: defer lock eval if freezing or frozen
We were only deferring if frozen.  But if freezing we need to too, because
of the way cap messages are deferred.  We defer cap messages if

 - inode is frozen
 - inode is freezing and locks are stable (to avoid starvation)

So if we are in a stable freezing state and start deferring caps, we can't
twiddle locks further or else we can

 - potentially starve (okay, in rare cases)
 - get stuck because we already started deferring cap messages

We would also screw up the cap message ordering if we became unstable again
and were allowed to start processing cap messages while others were still
deferred.

Signed-off-by: Sage Weil <sage.weil@dreamhost.com>