mds: rely on master to do anchor locks for slave_link_prep
The replica can't take all these locks without confusing things, since it
maybe need to unlock/relock, may screw up auth_pins, and worse.  The master
can take the locks.

The only problem is that the master may not know if the inode has already
been anchored if the lock hasn't cycled since then.  In that case, we take
more locks than we need to.

Signed-off-by: Sage Weil <sage.weil@dreamhost.com>