ProfLogger: don't start threads until it's safe
Introduce a new configuration variable, internal_safe_to_start_threads.
This will be set by common_init_finish once it is safe to start threads.
That will trigger callbacks for any configuration observers listening
for this event.

This is used by ProfLogger to hold off on starting the UNIX domain
socket thread until it is safe to do so.

Signed-off-by: Colin McCabe <colin.mccabe@dreamhost.com>