mds: track projected rename effect on subtree map
Renames can effect the subtree map.  We don't actually update it until the
rename commits, but while it is in flight to the jouranl we may journal
an ESubtreeMap event, which should reflect the change.

Adjust all callers (that journal) to project their change.

Mirror the logic in adjust_subtree_after_rename() in the subtree map
journaling code.  Be a bit careful, because we only journal subtrees that
we are auth for, so there are some additional checks to get a consistent
result.

Signed-off-by: Sage Weil <sage@newdream.net>