mds: make two passes on scatter_nudge
It's possible for scatter_nudge on a scatterlock in LOCK with dirty set to
go to MIX immediately and remain stable.  Give two 'nudge' passes before
we stop to avoid looping.

This fixes an assert failure where a nudge from log trimming ended up in a
stable state and asserted (!c).  The second pass will go trigger the dirty
writebehind.

Signed-off-by: Sage Weil <sage@newdream.net>