mds: mark ambig imports in ESubtreeMap during resolve
During resolve we may journal EImportFinish(true/false) as we resolve our
imports/exports.  And as a side-effect we may journal an ESubtreeMap.  We
need to properly mark ambig subtrees in that entry based on the
my_ambiguous_imports (resolve state), not just the migrator state (for the
active mds).

Note that the other Migrator::is_ambiguous_import() user
(send_resolve_now()) already does this correctly.

Signed-off-by: Sage Weil <sage.weil@dreamhost.com>