osd: Initialize new PGs with correct info.history.same_primary_since
Previously we were initializing based on the local osdmap epoch, which
is often correct, but if we process the MOSDPGCreate in an epoch after
the PG was created in the OSDMap, we could have problems with clients
sending out messages based on the creation epoch which the OSD would
reject as being earlier than same_primary_since. See bug #1357.

Signed-off-by: Greg Farnum <gregory.farnum@dreamhost.com>