osd: refactor PG creation slightly.
We want to carefully set up the PG History. In most cases this will
do the same thing as previously, since it's a brand new PG. However,
it's possible for the monitor to send out MOSDPGCreate messages for
the same PG to more than one OSD over many epochs if the OSDs are down
or are slow to respond. In that case we need to carefully track
history to make sure we don't lose old data.

Signed-off-by: Greg Farnum <gregory.farnum@dreamhost.com>