pgmon: create ALL pgs in new pools with correct creation epoch
5bb07df6aa4684ebd2e70437081dea170464d8ee tried to do this, but it
only set them properly for localized PGs. Whoops!
Additionally, we do NOT want to do this for new PGs in pre-existing
pools. Unfortunately, we have no way of guaranteeing that these new
PGs in old pools have the right epoch -- the data doesn't exist.
I'll discuss with other team members; it's possible that last_change
is in fact supposed to deal with this and simply doesn't.
Meanwhile, I've created a new bug to track this: #1365.

Signed-off-by: Greg Farnum <gregory.farnum@dreamhost.com>