pgmon: call check_osd_map via a new on_active implementation
Previously it was possible to lose PG creations if a monitor election
happened at the right time. The issue would get rectified on the
next OSDMap update, but that could take...a while. (My observed time
when I discovered the bug had it go without creation for 43 minutes,
at which point I killed it.)

Signed-off-by: Greg Farnum <gregory.farnum@dreamhost.com>