pgmon: use pool.get_last_change whenever creating new PGs
We maintain last_change properly now, so we can use it at any time.

It may still be possible that we can get PGs with the wrong epoch
created if, somehow, we do multiple expansions without
check_osd_map getting called between them. That's a pretty unlikely
occurrence, though, and I'm not sure that it's actually possible.

Signed-off-by: Greg Farnum <gregory.farnum@dreamhost.com>