objecter: allow requesting specific maps in maybe_request_map
Use this capability so that wait_for_new_map can specify a specific
map it wants go get. I am extending the sins of the fathers by
allowing a default value of 0 for the epoch here, but removing it
is even clumsier and there are lots of uses where it legitimately
doesn't care what epoch it gets.

Fixes #1383.

Signed-off-by: Greg Farnum <gregory.farnum@dreamhost.com>