mon: health not ok when up < in osds
We were warning if there were any not up or in osds.  Instead, warn if
there are any osds that are in but not up.  That means if a node fails
and successfully marks the node out and retracts onto remaining nodes, the
ceph cluster is healthy again.

Presumably the fact that the nodes failed should raise other alerts,
because those specific daemons/nodes are not healthy.

Signed-off-by: Sage Weil <sage@newdream.net>