client: only flush imported cap if it is now auth
We can get a cap import that isn't auth if there was a racing import
because the cap since migrated to another node.  We do not want to flush
the caps in that case.  Not to this mds, certainly, and not even to
another, because that would have happened when we got the other import.

Signed-off-by: Sage Weil <sage.weil@dreamhost.com>