cconf: do not common_init_finish
All that common_init_finish() does is indicate that we're done initializing
and we're allowed to start up extra threads and do the wonky things that
daemons like to do (like set up the admin socket).  Since cconf is just
examining the config, we don't want to do any of that.

Signed-off-by: Sage Weil <sage@newdream.net>